symlink issue

+++ kubux-wallpaper-generator.py
@@ -453,18 +453,20 @@ def download_image(url, file_name, prompt, error_callback=fallback_show_error):
 
     try:
         link_path=os.path.join(DOWNLOAD_DIR, file_name)
+        if os.path.lexists(link_path):
+            os.remove(link_path)
         os.symlink(save_path, link_path)
     except Exception as e:
-        os.remove(link_path)
         message = f"Failed to link image: {e}"
         error_callback("File system error,", message)
 
     try:
         link_path=os.path.join(IMAGE_DIR, file_name)
+        if os.path.lexists(link_path):
+            os.remove(link_path)
         os.symlink(save_path, link_path)
         return link_path
     except Exception as e:
-        os.remove(link_path)
         message = f"Failed to link image: {e}"
         error_callback("File system error,", message)
         return None
